+++ includes/config-system/class-schema.php
@@ -13,7 +13,6 @@
 
 namespace Newspack_Nodes\Config_System;
 
-
 \defined( 'ABSPATH' ) || exit;
 
 class Schema {